% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.incast_cv}
\alias{autoplot.incast_cv}
\title{Plot cross-validation model rankings}
\usage{
\method{autoplot}{incast_cv}(object, ...)
}
\arguments{
\item{object}{An \code{incast_cv} object returned by \code{\link{get_cv}}.}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Plot model performance across time series using the weighted interval score
(WIS). By default, models are ranked using relative WIS, where values below
1 indicate better-than-average performance and values above 1 indicate
worse-than-average performance. If only one model is present, raw WIS
values are displayed instead.
}
\details{
Each panel corresponds to a single time series.
}
