% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.incast_ncast}
\alias{autoplot.incast_ncast}
\title{Plot a nowcast}
\usage{
\method{autoplot}{incast_ncast}(object, ...)
}
\arguments{
\item{object}{An \code{incast_ncast} object returned by
\code{\link{get_ncast}}.}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Plot reported counts alongside nowcast distributions. Nowcasts are shown as
the median, with 50\% and 95\% credible intervals.
}
\details{
Each panel corresponds to a single time series.
}
