% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Fetch hospitalisation data}
\usage{
get_data(pathogen, geo_value, revisions = FALSE)
}
\arguments{
\item{pathogen}{One of "covid", "flu" or "rsv".}

\item{geo_value}{Geographic value(s) to fetch, as per
\link[epidatr]{pub_covidcast}.}

\item{revisions}{Logical. If \code{TRUE}, fetch the full revision history
needed by \code{\link{get_ncast}}. Default \code{FALSE} (latest only).}
}
\value{
An \code{incast_data} object (see \code{\link{check_data}}).
}
\description{
Fetch confirmed US hospital admissions (COVID-19, influenza or RSV) from
NHSN via \link[epidatr]{pub_covidcast}.
}
\examples{
get_data(pathogen = "covid", geo_value = "ny")

# Revision history for nowcasting
get_data(pathogen = "covid", geo_value = "ca", revisions = TRUE)
}
\author{
Cyril Geismar
}
