% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc_data.R
\docType{data}
\name{loc_data}
\alias{loc_data}
\title{Location Dataset}
\format{
A data frame with 53 rows and 12 variables:
\describe{
\item{abbreviation}{State abbreviation code}
\item{location}{State abbreviation code 2}
\item{location_name}{Full state name}
\item{population}{State population}
}
}
\source{
Where the data came from ??
}
\usage{
loc_data
}
\description{
US states census data. Includes state abbreviations, full state names, ????
}
\keyword{datasets}
