% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_respilens.R
\name{metadata_key}
\alias{metadata_key}
\title{Build RespiLens metadata key from \code{model_out_tbl} data}
\usage{
metadata_key(model_out_tbl)
}
\arguments{
\item{model_out_tbl}{Forecast tibble from \code{get_fcast()}.}
}
\value{
Named list for RespiLens metadata key
}
\description{
Build RespiLens metadata key from \code{model_out_tbl} data
}
