% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.accidda_cv}
\alias{print.accidda_cv}
\title{Print an \code{accidda_cv} object}
\usage{
\method{print}{accidda_cv}(x, ...)
}
\arguments{
\item{x}{An \code{accidda_cv} object.}

\item{...}{Ignored.}
}
\description{
Display a summary of the target, series, data window, and
cross-validation settings.
}
