% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.accidda_data}
\alias{print.accidda_data}
\title{Print an \code{accidda_data} object}
\usage{
\method{print}{accidda_data}(x, ...)
}
\arguments{
\item{x}{An \code{accidda_data} object.}

\item{...}{Ignored.}
}
\description{
Print an \code{accidda_data} object
}
