% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.incast_fcast}
\alias{print.incast_fcast}
\title{Print an \code{incast_fcast} object}
\usage{
\method{print}{incast_fcast}(x, ...)
}
\arguments{
\item{x}{An \code{incast_fcast} object.}

\item{...}{Ignored.}
}
\description{
Display a summary of the target, series, forecast period, and models used.
}
