% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.incast_ncast}
\alias{print.incast_ncast}
\title{Print an \code{incast_ncast} object}
\usage{
\method{print}{incast_ncast}(x, ...)
}
\arguments{
\item{x}{An \code{incast_ncast} object.}

\item{...}{Ignored.}
}
\description{
Display a summary of the target, series, data window, and nowcast period.
}
