% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R, R/autoplot.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tibble}
\alias{autoplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{as_tibble}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}}

