% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get hospitalisation data}
\usage{
get_data(pathogen, geo_value, revisions = FALSE)
}
\arguments{
\item{pathogen}{Character. One of "covid", "flu" or "rsv".}

\item{geo_value}{Character vector. Geographic value as per
\link[epidatr]{pub_covidcast} to filter by.}

\item{revisions}{Logical. If \code{TRUE}, fetches all available revision
history, producing data ready for \code{\link{get_ncast}}.
Default is \code{FALSE} (latest version only).}
}
\value{
An \code{accidda_data} object (see \code{\link{check_data}}).
}
\description{
Fetch confirmed US hospital admissions from the NHSN source
for COVID-19, influenza, or RSV using \link[epidatr]{pub_covidcast}.
}
\details{
By default the latest version of each observation is returned. Set
\code{revisions = TRUE} to retrieve all available revision history,
which is needed by \code{\link{get_ncast}}.
}
\examples{
get_data(pathogen = "covid", geo_value = "ny")

# Fetch revision history for nowcasting
get_data(pathogen = "covid", geo_value = "ca", revisions = TRUE)
}
\seealso{
\link[epidatr]{pub_covidcast}

\link[hubverse]{https://docs.hubverse.io/en/stable/user-guide/target-data.html}
}
