% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get hospitalisation data}
\usage{
get_data(pathogen, geo_values = "*")
}
\arguments{
\item{pathogen}{Character. One of "covid", "flu" or "rsv".}

\item{geo_values}{Character vector. Geographic values as per \link[epidatr]{pub_covidcast} to filter by. Defaults to "*" (all states).}
}
\value{
A tibble in \code{hubverse} format with columns:
\itemize{
\item \code{as_of}: Date when the data was issued
\item \code{location}: State code
\item \code{target}: Target description (e.g., "wk inc covid hosp")
\item \code{target_end_date}: End date of the target week
\item \code{observation}: Observed hospital admissions
}
}
\description{
Fetch the most recent confirmed US hospital admissions from the NHSN source
for COVID-19, influenza, or RSV using \link[epidatr]{pub_covidcast}.
}
\examples{
get_data(pathogen = "covid", geo_values = c("ca", "ny"))
}
\seealso{
\link[epidatr]{pub_covidcast}

\link[hubvserse]{https://docs.hubverse.io/en/stable/user-guide/target-data.html}
}
