% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.accidda_fcast}
\alias{print.accidda_fcast}
\title{Print an \code{accidda_fcast} object}
\usage{
\method{print}{accidda_fcast}(x, ...)
}
\arguments{
\item{x}{An \code{accidda_fcast} object.}

\item{...}{Ignored.}
}
\description{
Print an \code{accidda_fcast} object
}
