% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.accidda_ncast}
\alias{print.accidda_ncast}
\title{Print an \code{accidda_ncast} object}
\usage{
\method{print}{accidda_ncast}(x, ...)
}
\arguments{
\item{x}{An \code{accidda_ncast} object.}

\item{...}{Ignored.}
}
\description{
Print an \code{accidda_ncast} object
}
