% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.accidda_ncast}
\alias{print.accidda_ncast}
\title{Print an accidda_ncast object}
\usage{
\method{print}{accidda_ncast}(x, ...)
}
\arguments{
\item{x}{An object of class \code{accidda_ncast}.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Displays a concise summary of an \code{accidda_ncast} object.
}
